Adding GameLoader script
GameLoader builds a singleton to the WorldManager and the SoundManager.
Added WorldManager to manage the list of missions and the world selection
screen possibly. Added TileMap class to manage the list of Tiles
instead. This is an attempt to remove tile based responsibilities
from GameManager.

Deleted Game.cs. It wasn't needed. It was merely an example.