Adding TileGrid script and prefab
This script is the visual representation of a mission using
customly drawn meshes with materials made from images.

TODO: All of the materials must be fed into the script but I think
this can be modified via level somehow.
TODO: I think I can remove the need for all the materials by creating
a giant spritemap and adjusting the uv map to find the texture within
the material.